Remove hitch() as it isn't very useful
The goal of the hitch() function was to provide an easier way to maintain the typical prototype chain lookup properties of class methods. There were two main problems with this. First, it's not that hard to get this behavior when writing plain JavaScript. Second, the errors thrown when the prototype method is deleted were hard to trace back to the root problem. The default pattern of using an arrow function to reference the prototype's method would at least throw an error in the arrow function's definition, which is directly in the class declaration itself.

Here's the default approach using standard JavaScript:

```js
class Example {
  constructor(eventEmitter) {
    eventEmitter.addEventListener(this.handleMessage);
  }
  handleMessage = (...args) => this._handleMessage(...args);
  _handleMessage() { /* ... */ }
}
```